Split model page static data
